requirements.txt
# Core dependencies
python-docx==1.1.0
Flask==3.0.0
Werkzeug==3.0.1
Pillow>=9.3.0
requests>=2.28.0
matplotlib>=3.6.0

# Markdown processing
markdown2==2.4.12

# Optional dependencies
# pypandoc==1.13  # For pandoc conversion (optional)